+++ domains/Discussions/Database/Migrations/2020_10_19_175943_create_question_answers_table.php
@@ -1,0 +1,26 @@
+<?php
+
+use Domains\Accounts\Models\User;
+use Domains\Discussions\Models\Question;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateQuestionAnswersTable extends Migration
+{
+    public function up(): void
+    {
+        Schema::create('question_answers', function (Blueprint $table) {
+            $table->id();
+            $table->foreignIdFor(User::class, 'author_id');
+            $table->foreignIdFor(Question::class, 'question_id');
+            $table->text('content');
+            $table->timestampsTz();
+            $table->softDeletesTz();
+
+            $table->index('question_id');
+            $table->index('author_id');
+            $table->index('created_at');
+        });
+    }
+}