+++ domains/Discussions/routes.php
@@ -1,5 +1,6 @@
 <?php
 
+use Domains\Discussions\Controllers\AnswersStoreController;
 use Domains\Discussions\Controllers\QuestionsStoreController;
 use Domains\Discussions\Controllers\QuestionsUpdateController;
 use Illuminate\Support\Facades\Route;
@@ -14,4 +15,10 @@ Route::group(['middleware' => 'auth'], function () {
         'as' => 'questions.update',
         'uses' => QuestionsUpdateController::class,
     ]);
+
+    Route::post('/questions/{questionId}/answers', [
+        'as' => 'questions.answers',
+        'uses' => AnswersStoreController::class
+    ]);
 });
+