+++ domains/Discussions/routes.php
@@ -25,6 +25,5 @@ Route::group(['middleware' => 'auth'], function () {
 
 Route::get('questions/{questionId}', [
     'as' => 'questions.view',
-    'middleware' => 'throttle_guest:30,1',
     'uses' => QuestionsViewController::class
 ]);