+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to `laravel-portugal/api` will be documented in this file
 ### Changed
 
 - Updated Accounts' domain endpoints structure
+- Changed timestamps columns on links and tags tables (#28, #29)
 
 ### Deprecated