+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to `laravel-portugal/api` will be documented in this file
 - First version of the API documentation
 - A guest should be able to login and logout (#37)
 - Add account types and permissions (#42)
+- An authenticated user can post an answer to a question (#31)
 
 ### Changed