+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to `laravel-portugal/api` will be documented in this file
 
 ### Added
 
-- A guest or an authenticated user can see details of a question
+- A guest or an authenticated user can see details of a question (#48)
 
 ### Changed